mai lam tiep

+++ BT-Tuyen/02-11-2024/QUANTRONG.cpp
@@ -6,7 +6,7 @@ int main(){
     ll n;
     cin>>n;
     vector<int>a(n);
-    for(ll i = 0; i < n;i++)cin>>a[i];
+    for(int i = 0; i < n;i++)cin>>a[i];
     vector<pair<int, int>> pos;
     vector<int> danhdau(1e3+5, 0);
     vector<int> start(1e4+5, -1);
@@ -16,16 +16,16 @@ int main(){
         end[a[i]] = i;
         if(danhdau[a[i]] == 1) start[a[i]] = i;
     }
-    for(ll i = 0; i < n; i++){
+    for(int i = 0; i < n; i++){
         if(start[i] != -1 && end[i] != -1){
             pos.push_back({start[i], end[i]});
         }
     }
     ll maxx = 0;
-    for(ll i = 0; i < pos.size(); i++){
-        ll curr = 0;
+    for(int i = 0; i < pos.size(); i++){
+        int curr = 0;
         //cout<<pos[i].first<<" "<<pos[i].second<<endl;
-        for(ll j = pos[i].first; j <= pos[i].second; j++){
+        for(int j = pos[i].first; j <= pos[i].second; j++){
             curr += a[j];
         }
         if(curr > maxx) maxx = curr;
